refactor(mobile): extract symptom preset editor hook and components, add screen tests
